+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
+***Fixed:***
+
+- Prevent passing `TERM` environment variable to SSH connections
+
 ## 0.12.5 - 2025-05-12
 
 ***Fixed:***