+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
+## 0.13.0 - 2025-05-27
+
 ***Added:***
 
 - Pass local Git author details to developer environments